update paths and tailwind config
